requirements.txt
Flask==1.0.2
Flask-API==1.0
requests==2.20.0
gunicorn==19.9.0
python-dotenv==0.9.1
ipaddress==1.0.19
psycopg2-binary==2.7.6.1